Move type checking code to another file
